# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/getCityApi": {
            "get": {
                "tags": [
                    "getCityApi"
                ],
                "summary": "Get list of city",
                "description": "Returns list of projects",
                "operationId": "getCityApi",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/GetProductCategory": {
            "get": {
                "tags": [
                    "GetProductCategory"
                ],
                "summary": "Get list of city",
                "description": "Returns list of projects",
                "operationId": "GetProductCategory",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "api/RegisterManufacturerUser": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "RegisterManufacturerUser",
                "description": "RegisterManufacturerUser",
                "operationId": "RegisterManufacturerUser",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "NULL"
                                ],
                                "properties": {
                                    "telegram": {
                                        "type": "string",
                                        "format": "telegram",
                                        "example": "group_name"
                                    },
                                    "company_name": {
                                        "type": "string",
                                        "format": "name",
                                        "example": "justcode"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "format": "phone",
                                        "example": "93457898"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "111111"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "111111"
                                    },
                                    "individual_number": {
                                        "type": "text",
                                        "format": "text",
                                        "example": "7896847654321"
                                    },
                                    "watsap_phone": {
                                        "type": "text",
                                        "format": "text",
                                        "example": "+37493073584"
                                    },
                                    "i_agree": {
                                        "type": "string",
                                        "format": "I_agree",
                                        "example": "true"
                                    },
                                    "role_id": {
                                        "type": "string",
                                        "format": "role_id",
                                        "example": "3"
                                    },
                                    "made_in": {
                                        "type": "string",
                                        "format": "made_in",
                                        "example": "Armenia"
                                    },
                                    "price_of_metr": {
                                        "type": "string",
                                        "format": "price_of_metr",
                                        "example": "12"
                                    },
                                    "saite": {
                                        "type": "string",
                                        "format": "saite",
                                        "example": "www.facebook.com"
                                    },
                                    "show_room": {
                                        "type": "string",
                                        "format": "show_room",
                                        "example": "Да . нет. все"
                                    },
                                    "logo": {
                                        "type": "file",
                                        "format": "file",
                                        "example": "photo.png"
                                    },
                                    "percent_bonus[]": {
                                        "type": "text",
                                        "format": "text",
                                        "example": "10000^20000^9"
                                    },
                                    "sales_city[]": {
                                        "type": "text",
                                        "format": "text",
                                        "example": "1^москва"
                                    },
                                    "product_category[]": {
                                        "type": "text",
                                        "format": "text",
                                        "example": "1^зал"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "api/updateveryficode": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "updateveryficode",
                "description": "updateveryficode",
                "operationId": "updateveryficode",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "phone_veryfi_code": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "4587"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
    }
}